Split follows endpoint into followers and followings
