+++ app/views/api/v0/followers/index.json.jbuilder
