+++ app/views/api/v0/followings/organizations.json.jbuilder
