+++ app/views/api/v0/followings/tags.json.jbuilder
