+++ spec/requests/follows_api_spec.rb
@@ -30,52 +30,4 @@ RSpec.describe "FollowsApi", type: :request do
       expect(Follow.all.size).to eq(users_hash.size)
     end
   end
-
-  describe "GET /api/follows/followers" do
-    before do
-      sign_in user
-      user2.follow user
-      user.reload
-      get "/api/follows/followers"
-    end
-
-    it "returns followers list" do
-      expect(response.body).to include user2.name
-    end
-  end
-
-  describe "GET /api/follows/following" do
-    let(:tag) { create(:tag) }
-    let(:podcast) { create(:podcast) }
-    let(:organization) { create(:organization) }
-
-    before do
-      sign_in user
-      user.follow user2
-      user.follow tag
-      user.follow organization
-      user.follow podcast
-      user.reload
-    end
-
-    it "returns following users list" do
-      get "/api/follows/following_users"
-      expect(response.body).to include user2.name
-    end
-
-    it "returns following tag list" do
-      get "/api/follows/following_tags"
-      expect(response.body).to include tag.name
-    end
-
-    it "returns following organization list" do
-      get "/api/follows/following_organizations"
-      expect(response.body).to include organization.name
-    end
-
-    it "returns following podcast list" do
-      get "/api/follows/following_podcasts"
-      expect(response.body).to include podcast.name
-    end
-  end
 end
